chore: move components to dedicated folder
