+++ src/components/Loader/styles.tsx
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  frame: {
+    flex: 1,
+    width: '100%',
+  },
+  loader: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
