+++ src/components/Picker/components/EmojiCell.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { Text, TouchableHighlight } from 'react-native';
 
-import { useThemeContext } from '../../context/ThemeContext';
+import { useThemeContext } from '../../../context/ThemeContext';
 import styles from './styles';
 
 const EmojiCell = (props) => {
