+++ src/components/Picker/components/styles.js
