+++ src/components/SearchBar/index.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { TextInput, View } from 'react-native';
 
-import { useThemeContext } from '../context/ThemeContext';
+import { useThemeContext } from '../../context/ThemeContext';
 import styles from './styles';
 
 const SearchBar = (props) => {
