+++ src/components/TabBar/styles.js
