+++ src/components/index.tsx
@@ -1,3 +1,4 @@
+export { default as Loading } from './Loader';
 export { default as Picker } from './Picker';
 export { default as SearchBar } from './SearchBar';
 export { default as TabBar } from './TabBar';
